fix some error

+++ src/smartban/manager/BanManager.java
@@ -187,7 +187,7 @@ public class BanManager {
 		}
 		for (Entry<String, Object> entry : db.getDB("banlist").getAll().entrySet()) {
 			Map<String, Object> map = (Map<String, Object>) entry.getValue();
-			if (getIP(player).equals(map.get("ip")) || getClientId(player) == (long) map.get("clientid")) {
+			if (getIP(player).equals(map.get("ip")) || getClientId(player) == (double) map.get("clientid")) {
 				return entry.getKey();
 			}
 		}
